Updated airport system for flight planning
Split flight planning view model into partial files
